apps/api/pyproject.toml
[project]
name = "zero-employee-orchestrator"
version = "0.1.7"
description = "Zero-Employee Orchestrator – AI Orchestration Platform for multi-agent business automation with human approval and full auditability"
requires-python = ">=3.11"
license = "MIT"
keywords = [
    "ai",
    "orchestrator",
    "multi-agent",
    "automation",
    "llm",
    "fastapi",
    "ollama",
    "business-automation",
]
authors = [
    { name = "OrosiTororo" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.7",
    "litellm>=1.60,!=1.82.7,!=1.82.8",
    "aiohttp>=3.13.4",
    "python-jose[cryptography]>=3.3",
    "cryptography>=46.0.6",
    "bcrypt>=4.0",
    "slowapi>=0.1.9",
    "httpx>=0.28",
    "websockets>=12.0",
    "apscheduler>=3.10",
    "structlog>=24.0",
    "python-multipart>=0.0.9",
    "packaging>=23.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
monitoring = [
    "sentry-sdk[fastapi]>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.urls]
Homepage = "https://github.com/OrosiTororo/Zero-Employee-Orchestrator"
Repository = "https://github.com/OrosiTororo/Zero-Employee-Orchestrator"
Issues = "https://github.com/OrosiTororo/Zero-Employee-Orchestrator/issues"
Documentation = "https://github.com/OrosiTororo/Zero-Employee-Orchestrator/tree/master/docs"

[project.scripts]
zero-employee = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "B008",
    "UP042",
    "B904",
    "SIM105",
    "SIM102",
    "SIM117",
    "SIM108",
    "SIM110",
    "UP046",
]

[tool.ruff.lint.per-file-ignores]
# Existing modules that carry pre-v0.1.8 line-length debt (Japanese-heavy
# strings, large inline manifests, multi-line f-strings). E501 is enforced on
# all other files — new code must stay within 100 columns. Revisit and shrink
# this list as each file is split / wrapped in follow-up refactors.
"app/banner.py" = ["E501"]
"app/cli.py" = ["E501"]
"app/core/config.py" = ["E501"]
"app/api/routes/registry.py" = ["E501"]
"app/api/routes/security_settings.py" = ["E501"]
"app/orchestration/executor.py" = ["E501"]
"app/orchestration/meta_skills.py" = ["E501"]
"app/security/redteam.py" = ["E501"]
"app/security/prompt_guard.py" = ["E501"]
"app/services/nl_command_service.py" = ["E501"]
"app/services/plugin_loader.py" = ["E501"]
"app/services/registry_service.py" = ["E501"]
"app/services/self_improvement_service.py" = ["E501"]
# The v0.1.7 self-improvement package split preserves the original string
# literals (long f-string prompts, Japanese log messages). Cleaning those up
# is a separate pass.
"app/services/self_improvement/skill_analyzer.py" = ["E501"]
"app/services/self_improvement/skill_improver.py" = ["E501"]
"app/services/self_improvement/judge_tuner.py" = ["E501"]
"app/services/self_improvement/failure_to_skill.py" = ["E501"]
"app/services/self_improvement/ab_test.py" = ["E501"]
"app/services/self_improvement/test_generator.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.mypy]
python_version = "3.12"
strict = true